requirements.txt
requests
python-dotenv
slack-sdk
google-api-python-client
google-auth
anthropic
langchain
ipython
langgraph
langchain-google-genai
langchain-communitymcp
